+++ lib/ui/chat/chat_info/edit_group_screen.dart
@@ -192,6 +192,7 @@ class _EditGroupScreenState extends ConsumerState<EditGroupScreen> {
                             displayName: group.name,
                             size: 96.w,
                             showBorder: true,
+                            pubkey: group.nostrGroupId,
                           ),
                         ),
                         Gap(36.h),