+++ lib/ui/contact_list/group_chat_details_sheet.dart
@@ -4,6 +4,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 import 'package:logging/logging.dart';
+import 'package:whitenoise/config/providers/avatar_color_provider.dart';
 import 'package:whitenoise/config/providers/create_group_provider.dart';
 import 'package:whitenoise/config/states/create_group_state.dart';
 import 'package:whitenoise/domain/models/contact_model.dart';
@@ -55,11 +56,15 @@ class _GroupChatDetailsSheetState extends ConsumerState<GroupChatDetailsSheet> w
   final TextEditingController _groupNameController = TextEditingController();
   final TextEditingController _groupDescriptionController = TextEditingController();
   Group? createdGroup;
+  Color? _previewColor;
+
   @override
   void initState() {
     super.initState();
     _groupNameController.addListener(_onGroupNameChanged);
     _groupDescriptionController.addListener(_onGroupDescriptionChanged);
+    _previewColor = ref.read(avatarColorProvider.notifier).generateRandomColor();
+
     WidgetsBinding.instance.addPostFrameCallback((_) {
       ref.read(createGroupProvider.notifier).filterContactsWithKeyPackage(widget.selectedContacts);
     });
@@ -81,10 +86,22 @@ class _GroupChatDetailsSheetState extends ConsumerState<GroupChatDetailsSheet> w
     await ref
         .read(createGroupProvider.notifier)
         .createGroup(
-          onGroupCreated: (createdGroup) {
+          onGroupCreated: (createdGroup) async {
             if (createdGroup != null && mounted) {
+              final previewColor = _previewColor;
+              if (previewColor != null) {
+                await ref
+                    .read(avatarColorProvider.notifier)
+                    .setColorDirectly(
+                      createdGroup.nostrGroupId,
+                      previewColor,
+                    );
+              }
+
               this.createdGroup = createdGroup;
-              context.pop();
+              if (mounted) {
+                context.pop();
+              }
             }
           },
         );
@@ -165,6 +182,7 @@ class _GroupChatDetailsSheetState extends ConsumerState<GroupChatDetailsSheet> w
                       displayName: displayName,
                       size: 96.w,
                       showBorder: true,
+                      color: _previewColor,
                     );
                   },
                 ),
@@ -261,6 +279,7 @@ class _GroupChatDetailsSheetState extends ConsumerState<GroupChatDetailsSheet> w
                               displayName: contact.displayName,
                               size: 30.w,
                               showBorder: true,
+                              pubkey: contact.publicKey,
                             ),
                             Gap(8.w),
                             SizedBox(