+++ lib/ui/settings/profile/edit_profile_screen.dart
@@ -5,6 +5,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 import 'package:whitenoise/config/extensions/toast_extension.dart';
+import 'package:whitenoise/config/providers/active_account_provider.dart';
 import 'package:whitenoise/config/providers/edit_profile_screen_provider.dart';
 import 'package:whitenoise/config/states/profile_state.dart';
 import 'package:whitenoise/ui/core/themes/assets.dart';
@@ -158,6 +159,12 @@ class _ProfileState extends ConsumerState<EditProfileScreen> {
                                             displayName: displayName,
                                             size: 96.w,
                                             showBorder: imageUrl.isEmpty,
+                                            pubkey:
+                                                ref
+                                                    .watch(activeAccountProvider)
+                                                    .value
+                                                    ?.account
+                                                    ?.pubkey,
                                           );
                                         },
                                       ),