+++ lib/ui/settings/profile/share_profile_screen.dart
@@ -102,6 +102,7 @@ class _ShareProfileScreenState extends ConsumerState<ShareProfileScreen> {
                               displayName: profile?.displayName ?? '',
                               size: 96.w,
                               showBorder: true,
+                              pubkey: state.account?.pubkey,
                             ),
                             Gap(8.h),
                             Text(